feat: Adds a github formatter (#463)
* refactor: Extracts formatter modules

* refactor: Uses a behaviour over string and atom matching

* feat: Adds a github formatter

* chore: Updates log level and documentation

* fix: Cleans up some warnings
